This adds cherrypy config and main function
