+++ uniqueid/__init__.py
@@ -1,2 +1,40 @@
 #!/usr/bin/python
-"""Unique ID module."""
+"""UniqueID Module."""
+from argparse import ArgumentParser
+from json import dumps
+import cherrypy
+from uniqueid.globals import CHERRYPY_CONFIG
+from uniqueid.rest import Root
+from uniqueid.orm import create_tables
+
+
+def error_page_default(**kwargs):
+    """The default error page should always enforce json."""
+    cherrypy.response.headers['Content-Type'] = 'application/json'
+    return dumps({
+        'status': kwargs['status'],
+        'message': kwargs['message'],
+        'traceback': kwargs['traceback'],
+        'version': kwargs['version']
+    })
+
+
+def main():
+    """Main method to start the httpd server."""
+    parser = ArgumentParser(description='Run the uniqueid server.')
+    parser.add_argument('-c', '--config', metavar='CONFIG', type=str,
+                        default=CHERRYPY_CONFIG, dest='config',
+                        help='cherrypy config file')
+    parser.add_argument('-p', '--port', metavar='PORT', type=int,
+                        default=8051, dest='port',
+                        help='port to listen on')
+    parser.add_argument('-a', '--address', metavar='ADDRESS',
+                        default='localhost', dest='address',
+                        help='address to listen on')
+    args = parser.parse_args()
+    create_tables()
+    cherrypy.config.update({
+        'server.socket_host': args.address,
+        'server.socket_port': args.port
+    })
+    cherrypy.quickstart(Root(), '/', args.config)
